pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "qabda"
version = "0.1.0"
description = "Quantum-Assisted Big Data Analytics (QABDA): hybrid Spark + quantum subroutines for ultra-high-dimensional analytics."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Vishal Bharadwaj Meruga"}]
keywords = ["quantum", "big-data", "spark", "qiskit", "pennylane", "dimensionality-reduction", "vqa", "qaoa"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "typer>=0.9",
  "pydantic>=2.0",
  "rich>=13.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
spark = ["pyspark>=3.4"]
qiskit = ["qiskit>=1.0", "qiskit-aer>=0.14"]
pennylane = ["pennylane>=0.35"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.5", "mypy>=1.7", "pre-commit>=3.5"]

[project.scripts]
qabda = "qabda.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
